@@ -6,7 +6,7 @@
   "packageManager": "yarn@4.0.2",
   "typings": "index.d.ts",
   "license": "MIT",
-  "version": "4.0.0-beta.119",
+  "version": "4.0.0-beta.120",
   "author": {
     "name": "kobezzza",
     "email": "kobezzza@gmail.com",
@@ -53,11 +53,11 @@
     "@babel/preset-env": "7.16.11",
     "@babel/preset-typescript": "7.16.7",
     "@config/config": "npm:config@1.31.0",
-    "@swc/core": "1.2.153",
-    "@swc/helpers": "0.3.8",
+    "@swc/helpers": "0.5.10",
     "@vue/compiler-sfc": "3.2.47",
     "chalk": "4.1.2",
-    "cookiejar": "^2.1.4",
+    "cookiejar": "2.1.4",
+    "core-js": "3.37.0",
     "cross-env": "7.0.3",
     "dompurify-v2": "npm:dompurify@2.4.7",
     "dompurify-v3": "npm:dompurify@3.0.8",
@@ -76,12 +76,13 @@
     "@pzlr/build-core": "2.13.0",
     "@pzlr/stylus-inheritance": "3.3.0",
     "@statoscope/webpack-plugin": "5.25.1",
+    "@swc/core": "1.4.14",
     "@types/webpack-env": "1.18.1",
     "@v4fire/config": "1.0.0",
     "@v4fire/design-system": "1.23.0",
     "arg": "5.0.1",
     "autoprefixer": "10.4.2",
-    "browserslist": "4.19.1",
+    "browserslist": "4.23.0",
     "buble": "0.20.0",
     "cli-progress": "3.10.0",
     "collection.js": "6.8.1",
@@ -137,16 +138,17 @@
     "svgo": "2.8.0",
     "svgo-loader": "3.0.0",
     "svgo-sync": "0.5.1",
+    "swc-loader": "0.2.6",
     "terser": "5.31.0",
-    "terser-webpack-plugin": "5.3.1",
-    "to-string-loader": "^1.2.0",
-    "ts-loader": "9.2.6",
+    "terser-webpack-plugin": "5.3.10",
+    "to-string-loader": "1.2.0",
+    "ts-loader": "9.5.1",
     "typescript": "4.6.2",
     "typograf": "6.14.0",
     "upath": "2.0.1",
     "url-loader": "4.1.1",
-    "webpack": "5.82.1",
-    "webpack-cli": "5.0.1"
+    "webpack": "5.91.0",
+    "webpack-cli": "5.1.4"
   },
   "devDependencies": {
     "@playwright/test": "1.32.1",
@@ -155,14 +157,14 @@
     "@storybook/addon-links": "7.0.23",
     "@storybook/blocks": "7.0.23",
     "@storybook/testing-library": "0.0.14-next.2",
-    "@types/cookiejar": "^2",
+    "@types/cookiejar": "2",
     "@types/dompurify": "3.0.5",
     "@types/express": "4",
     "@types/jest": "29.5.1",
     "@types/jsdom": "16.2.15",
     "@types/semver": "7.3.10",
     "@types/webpack": "5.28.0",
-    "@v4fire/cli": "^1.6.0",
+    "@v4fire/cli": "1.6.0",
     "@v4fire/core": "github:V4Fire/Core#v4",
     "@v4fire/linters": "git+https://github.com/v4fire/linters#rework_rules",
     "@v4fire/storybook": "0.8.0",
@@ -186,9 +188,9 @@
   },
   "resolutions": {
     "sharp": "0.31.3",
-    "webpack": "5.82.1",
     "html-webpack-plugin": "5.5.0",
     "globals": "13.19.0",
+    "path-to-regexp": "3.2.0",
     "file-system-cache": "2.3.0"
   }
 }
